refactor: reorganize configuration
Copy [omerxx](https://github.com/omerxx/dotfiles/tree/master/nvim)'s
example to _hopefully_ simplify neovim configuration.
