--- lua/custom/plugins/formatter.lua
@@ -1,20 +1,0 @@
-return {
-    "stevearc/conform.nvim",
-    -- optional = true,
-    opts = {
-        formatters_by_ft = {
-            cs = { "csharpier" },
-            html = { { "prettierd", "prettier" } },
-            javascript = { { "biome", "prettierd", "prettier" } },
-            ruby = { "standardrb" },
-            rust = { "rustfmt" },
-            ["_"] = { "trim_whitespace" },
-        },
-        formatters = {
-            csharpier = {
-                command = "dotnet-csharpier",
-                args = { "--write-stdout" },
-            },
-        },
-    },
-}
