--- lua/custom/plugins/indent-blankline.lua
@@ -1,12 +1,0 @@
--- Show indentation
-return {
-    {
-        'lukas-reineke/indent-blankline.nvim',
-        main = 'ibl',
-        opts = {
-            indent = {
-                char = '┊',
-            },
-        },
-    },
-}
