--- lua/custom/plugins/treesitter.lua
@@ -1,36 +1,0 @@
-return {
-    {
-        -- Highlight, edit, and navigate code
-        'nvim-treesitter/nvim-treesitter',
-        build = function()
-            local ts_update = require('nvim-treesitter.install').update({ with_sync = true })
-            ts_update()
-        end,
-        config = function()
-            require'nvim-treesitter.configs'.setup {
-                auto_install = true,
-                disable = { 'markdown' },
-                ensure_installed = { 'lua', 'vim', 'ruby' },
-                highlight = {
-                    enable = true,
-                    additional_vim_regex_highlighting = { 'text', 'mail', 'markdown' },
-                },
-                incremental_selection = {
-                    enable = true,
-                    keymaps = {
-                        init_selection = "gnn",
-                        node_incremental = "grn",
-                        scope_incremental = "grc",
-                        node_decremental = "grm",
-                    },
-                },
-                indent = {
-                    enable = true,
-                },
-            }
-        end,
-    },
-
-    -- Additional text objects via treesitter
-    { 'nvim-treesitter/nvim-treesitter-textobjects' },
-}
