--- lua/custom/plugins/trouble.lua
@@ -1,9 +1,0 @@
-return {
-    {
-        'folke/trouble.nvim',
-        opts = {
-            mode = 'document_diagnostics',
-        },
-        cmd = 'TroubleToggle',
-    },
-}
