--- lua/custom/plugins/ufo.lua
@@ -1,47 +1,0 @@
-return {
-    {
-        'kevinhwang91/nvim-ufo',
-        dependencies = { 'kevinhwang91/promise-async' },
-        opts = {
-            provider_selector = function(bufnr, filetype, buftype)
-                return { 'treesitter', 'indent' }
-            end,
-            fold_virt_text_handler = function(virtText, lnum, endLnum, width, truncate)
-                local newVirtText = {}
-                local suffix = (' 󰁂 %d '):format(endLnum - lnum)
-                local sufWidth = vim.fn.strdisplaywidth(suffix)
-                local targetWidth = width - sufWidth
-                local curWidth = 0
-                for _, chunk in ipairs(virtText) do
-                    local chunkText = chunk[1]
-                    local chunkWidth = vim.fn.strdisplaywidth(chunkText)
-                    if targetWidth > curWidth + chunkWidth then
-                        table.insert(newVirtText, chunk)
-                    else
-                        chunkText = truncate(chunkText, targetWidth - curWidth)
-                        local hlGroup = chunk[2]
-                        table.insert(newVirtText, {chunkText, hlGroup})
-                        chunkWidth = vim.fn.strdisplaywidth(chunkText)
-                        -- str width returned from truncate() may less than 2nd argument, need padding
-                        if curWidth + chunkWidth < targetWidth then
-                            suffix = suffix .. (' '):rep(targetWidth - curWidth - chunkWidth)
-                        end
-                        break
-                    end
-                    curWidth = curWidth + chunkWidth
-                end
-                table.insert(newVirtText, {suffix, 'MoreMsg'})
-                return newVirtText
-            end,
-        },
-        init = function()
-            -- Using ufo provider need a large value, feel free to decrease the value
-            vim.opt.foldlevel = 99
-            vim.opt.foldlevelstart = 99
-            vim.opt.foldenable = true
-
-            vim.keymap.set('n', 'zR', function() require('ufo').openAllFolds() end)
-            vim.keymap.set('n', 'zM', function() require('ufo').closeAllFolds() end)
-        end,
-    },
-}
